### 2020-11-02 #2811 @whedon
[PRE REVIEW]: c-lasso: a Python package for constrained sparse regression and classification
**Submitting author:** @Leo-Simpson (<a href="http://orcid.org/0000-0003-0717-5244">Leo Simpson</a>)
**Repository:** <a href="https://github.com/Leo-Simpson/c-lasso" target ="_blank">https://github.com/Leo-Simpson/c-lasso</a>
**Version:** v1.0
