+++ paper/paper.bib
@@ -123,6 +123,7 @@ year = {2020}
   lccn={2008941148},
   series={Springer series in statistics},
   url={https://books.google.fr/books?id=eBSgoAEACAAJ},
+  doi={10.1007/bf02985802},
   year={2009},
   publisher={Springer}
 }
@@ -138,7 +139,7 @@ year = {2020}
 
 @article{Jeon:2020,
 author = {Jeon, Jong-June and Kim, Yongdai and Won, Sungho and Choi, Hosik},
-doi = {https://doi.org/10.1016/j.csda.2020.106958},
+doi = {10.1016/j.csda.2020.106958},
 issn = {0167-9473},
 journal = {Computational Statistics {\&} Data Analysis},
 keywords = { Constraint, Microbiome data, Solution path algorithm,Penalized regression},
