### 2020-11-11 2a62678
use misclassification rate instead of hinge for CV in classification
