+++ classo/solve_R3.py
@@ -158,7 +158,7 @@ def pathlasso_R3(pb, path, n_active=False):
         pb.init = X[1]
 
         if (
-            sum([(abs(X[0][i]) > 1e-1) for i in range(len(X[0]))]) >= n_act
+            sum([(abs(X[0][i]) > 1e-5) for i in range(len(X[0]))]) >= n_act
             or type(X[1]) == str
         ):
             pb.init = save_init
