+++ classo/solve_R2.py
@@ -200,7 +200,7 @@ def pathlasso_R2(pb, path, n_active=False):
         BETA.append(X[0])
         pb.init = X[1]
         if (
-            sum([(abs(X[0][i]) > 1e-1) for i in range(len(X[0]))]) >= n_act
+            sum([(abs(X[0][i]) > 1e-5) for i in range(len(X[0]))]) >= n_act
             or type(X[1]) == str
         ):
             pb.init = save_init
