+++ README.md
@@ -1,4 +1,4 @@
-![arXiv](https://img.shields.io/badge/arXiv-<INDEX>-<COLOR>.svg)](https://arxiv.org/abs/2011.00898)
+[![arXiv](https://img.shields.io/badge/arXiv-<INDEX>-<COLOR>.svg)(https://arxiv.org/abs/2011.00898)
  
 <img src="https://i.imgur.com/2nGwlux.png" alt="c-lasso" height="150" align="right"/>
 
