### 2020-11-18 #2844 @whedon
[REVIEW]: c-lasso: a Python package for constrained sparse regression and classification
**Submitting author:** @Leo-Simpson (<a href="http://orcid.org/0000-0003-0717-5244">Leo Simpson</a>)
**Repository:** <a href="https://github.com/Leo-Simpson/c-lasso" target ="_blank">https://github.com/Leo-Simpson/c-lasso</a>
**Version:** v1.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.13589585.v1" target="_blank">10.6084/m9.figshare.13589585.v1</a>
