### 2020-11-18 Leo-Simpson/c-lasso#2 @jbytecode
[JOSS Review]: License 
Dear authors,

I am writing here due to the JOSS submission in [here](https://github.com/openjournals/joss-reviews/issues/2844).

It seems the software repository does not include a valid license. Please pick the proper one from the [list](https://opensource.org/licenses/alphabetical) (for example MIT) and place a license file using the [page](https://github.com/Leo-Simpson/c-lasso/community) in the repository.

Thank you in advance.
