### 2020-11-18 leo-simpson/classo#2 @muellsen
Thanks! We added an MIT license!
