### 2020-11-18 Leo-Simpson/c-lasso#4 @jbytecode
[JOSS Review]: Automated tests
Dear authors,

Due to reviewing process in [JOSS](https://github.com/openjournals/joss-reviews/issues/2844):

There is a section

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

in the reviewing list, however, it seems the software does not implement tests (at least unit tests). Please have a read on the unit testing in Python (for example [here](https://docs.python.org/3/library/unittest.html)) and implement tests for all of the functionality in the package.

Please add a `build and test` action in Actions part of the repository in [here](https://github.com/Leo-Simpson/c-lasso/actions) so when a commit is pushed, a triggered task will produce a report. You can also put a "PASSED" badge in your README.md file. Hence, the people can see whether the latest release works or not.

Thank you in advance.
