### 2020-11-18 leo-simpson/classo#4 @Leo-Simpson
Dear @jbytecode, thank you for your comment. I am currently working on that. I will tell you once it is ready.

Best regards.
