### 2020-11-18 Leo-Simpson/c-lasso#5 @jbytecode
[JOSS Review]: Quality of writing
Dear authors,

Due to the paper submission in [JOSS](https://github.com/openjournals/joss-reviews/issues/2844):

There is a section called

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

in the reviewing list.

*First of all, I am sorry if I open too much issues, however, it is very convenient to talk each single item in a separate window.*

The language of the paper is good. However expressions and statements are not well structured.  Please consider using a single (and unique) rule when you present code snippets in the manuscript. I am suggesting the general rule (if you don't adopt your own):

-  Please remove space before a comma, put a space after a comma like `1, 2, 3, ..., n`
- Please put spaces before and after an operator like `a = 5` and `2 + 2 = 4`
- Do not use spaces before and after a parenthesis in function calls

so the statement

`n,d,d_nonzero,k,sigma =100,100,5,1,0.5`

becomes

`n, d, d_nonzero, k, sigma = 100, 100, 5, 1, 0.5`

and the statement

`(X,C,y),sol = random_data(n,d,d_nonzero,k,sigma,zerosum=True, seed = 123 )`

becomes

`(X, C, y), sol = random_data(n, d, d_nonzero, k, sigma, zerosum = True, seed = 123)`

You can also consider saving space by refactoring

`print("Relevant variables  : {}".format(list(numpy.nonzero(sol)) ) )`

to

`print("Relevant variables: {}".format(list(numpy.nonzero(sol))))`

Please check all the codes in the manuscript.

Thank you in advance.
