### 2020-11-18 3b7d5b2
add logscale to the paths
