### 2020-11-19 c137593
correct code style in the examples
