+++ examples/Synthetic data/exampleJOSS.py
@@ -1,15 +1,15 @@
 # here is the path of the directory where one wants to save its figures
-path = '/Users/lsimpson/Desktop/GitHub/c-lasso/paper/figures/'
+path = "/Users/lsimpson/Desktop/GitHub/c-lasso/paper/figures/"
 import numpy
 from classo import classo_problem, random_data
 
-n,d,d_nonzero,k,sigma =100,100,5,1,0.5
-(X,C,y),sol = random_data(n,d,d_nonzero,k,sigma,zerosum=True, seed = 123 )
-print("Relevant variables  : {}".format(numpy.nonzero(sol)[0] ) )
+n, d, d_nonzero, k, sigma = 100, 100, 5, 1, 0.5
+(X, C, y), sol = random_data(n, d, d_nonzero, k, sigma, zerosum=True, seed=123)
+print("Relevant variables  : {}".format(numpy.nonzero(sol)[0]))
 
-problem  = classo_problem(X,y,C)
+problem = classo_problem(X, y, C)
 
-problem.formulation.huber  = True
+problem.formulation.huber = True
 problem.formulation.concomitant = False
 problem.formulation.rho = 1.5
 
@@ -21,17 +21,22 @@ problem.model_selection.LAMfixedparameters.lam = 0.1
 problem.solve()
 
 
-
-
-
-problem.solution.StabSel.save1 = path+problem.model_selection.StabSelparameters.numerical_method+'-StabSel'
-problem.solution.StabSel.save2 = path+problem.model_selection.StabSelparameters.numerical_method+'-StabSel-path'
-problem.solution.StabSel.save3 = path+problem.model_selection.StabSelparameters.numerical_method+'-StabSel-beta'
-#problem.solution.CV.save = path+'CV-beta'
-problem.solution.LAMfixed.save = path+problem.model_selection.PATHparameters.numerical_method+'-LAM-beta'
-problem.solution.PATH.save = path+problem.model_selection.LAMfixedparameters.numerical_method+'-PATH'
-
-
+problem.solution.StabSel.save1 = (
+    path + problem.model_selection.StabSelparameters.numerical_method + "-StabSel"
+)
+problem.solution.StabSel.save2 = (
+    path + problem.model_selection.StabSelparameters.numerical_method + "-StabSel-path"
+)
+problem.solution.StabSel.save3 = (
+    path + problem.model_selection.StabSelparameters.numerical_method + "-StabSel-beta"
+)
+# problem.solution.CV.save = path+'CV-beta'
+problem.solution.LAMfixed.save = (
+    path + problem.model_selection.PATHparameters.numerical_method + "-LAM-beta"
+)
+problem.solution.PATH.save = (
+    path + problem.model_selection.LAMfixedparameters.numerical_method + "-PATH"
+)
 
 
 print(problem.solution)
