### 2020-11-19 leo-simpson/classo#5 @Leo-Simpson
I think it is done now.

Best regards.
