+++ .github/workflows/python-publish.yml
@@ -22,10 +22,17 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install setuptools wheel twine
+        pip install --upgrade matplotlib
+        pip install --upgrade numpy
+        pip install --upgrade scipy
+        pip install --upgrade pandas
+        pip install --upgrade h5py
+
     - name: Build and publish
       env:
         TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
         TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
       run: |
+        rm -r dist
         python setup.py sdist bdist_wheel
         twine upload dist/*
