+++ .github/workflows/python-publish.yml
@@ -12,21 +12,6 @@ jobs:
 
     runs-on: ubuntu-latest
 
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: '3.x'
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install setuptools wheel twine
-        pip install --upgrade matplotlib
-        pip install --upgrade numpy
-        pip install --upgrade scipy
-        pip install --upgrade pandas
-        pip install --upgrade h5py
 
     - name: Build and publish
       env:
