+++ requirements.txt
@@ -1,0 +1,198 @@
+alabaster==0.7.12
+alembic==1.3.1
+altgraph==0.17
+anaconda-clean==1.0
+anaconda-client==1.7.2
+anaconda-navigator==1.9.12
+appdirs==1.4.4
+appnope==0.1.0
+argon2-cffi @ file:///opt/concourse/worker/volumes/live/b46b812d-eebc-4040-75c0-7fc24d9bfd54/volume/argon2-cffi_1596828479252/work
+asciitree==0.3.3
+async-generator==1.10
+attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
+Babel==2.8.0
+backcall==0.2.0
+beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1601924105527/work
+black==20.8b1
+bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
+brotlipy==0.7.0
+c-lasso==1.0.0.9
+certifi==2020.6.20
+certipy==0.1.3
+cffi==1.14.0
+chardet==3.0.4
+click==7.1.2
+clyent==1.2.2
+conda==4.9.1
+conda-build==3.15.1
+conda-package-handling==1.7.0+0.g7c4a471.dirty
+cryptography @ file:///opt/concourse/worker/volumes/live/c92b6573-5445-4a21-4a76-9af93cc67a5e/volume/cryptography_1601046836410/work
+cvxopt==1.2.0
+cvxpy==1.0.25
+cx-Freeze==6.2
+cycler==0.10.0
+dataclasses==0.7
+decorator==4.4.2
+defusedxml==0.6.0
+dill==0.3.1.1
+docutils==0.16
+ecos==2.0.7.post1
+entrypoints==0.3
+et-xmlfile==1.0.1
+fasteners==0.15
+filelock==3.0.12
+future==0.18.2
+geographiclib==1.50
+geopy==2.0.0
+glob2==0.7
+h5py==2.10.0
+idna @ file:///tmp/build/80754af9/idna_1593446292537/work
+imagesize==1.2.0
+importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
+iniconfig==1.1.1
+ipykernel @ file:///opt/concourse/worker/volumes/live/43eaf185-e9cd-4e78-5afa-c19accf1339d/volume/ipykernel_1596206683810/work/dist/ipykernel-5.3.4-py3-none-any.whl
+ipyparallel==6.2.4
+ipython @ file:///opt/concourse/worker/volumes/live/8c453bb6-500b-4747-778c-a236ae4a574d/volume/ipython_1593447395779/work
+ipython-genutils==0.2.0
+ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
+jdcal==1.4.1
+jedi @ file:///opt/concourse/worker/volumes/live/565dbe5a-443f-465a-7ff2-b61f4a5f081d/volume/jedi_1598371617578/work
+Jinja2==2.11.2
+joblib==0.16.0
+json5==0.9.5
+jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
+jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
+jupyter-contrib-core==0.3.3
+jupyter-contrib-nbextensions==0.5.1
+jupyter-core==4.6.3
+jupyter-highlight-selected-word==0.2.0
+jupyter-latex-envs==1.4.6
+jupyter-nbextensions-configurator==0.4.1
+jupyterhub==1.0.0
+jupyterlab==2.2.6
[…]
