### 2020-11-19 cf2b3c4
Update example_COMBO.py
