### 2020-11-19 3c26f34
Update setup.cfg
