+++ setup.cfg
@@ -1,4 +1,7 @@
-[egg_info]
-tag_build = 
-tag_date = 0
-
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = classo/_version.py
+versionfile_build = classo/_version.py
+tag_prefix =
+parentdir_prefix = c-lasso
