### 2020-11-19 a516f55
Update setup.cfg
