### 2020-11-19 8281d58
Update setup.cfg
