+++ setup.cfg
@@ -4,4 +4,4 @@ style = pep440
 versionfile_source = classo/_version.py
 versionfile_build = classo/_version.py
 tag_prefix =
-parentdir_prefix = c-lasso/
+parentdir_prefix = c-lasso-
