### 2020-11-19 7ae1fef
Update setup.cfg
