+++ setup.cfg
@@ -1,22 +1,3 @@
-<<<<<<< HEAD
-[egg_info]
-tag_build = 
-tag_date = 0
-
-
-# See the docstring in versioneer.py for instructions. Note that you must
-# re-run 'versioneer.py setup' after changing this section, and commit the
-# resulting files.
-
-[versioneer]
-#VCS = git
-#style = pep440
-#versionfile_source =
-#versionfile_build =
-#tag_prefix =
-#parentdir_prefix =
-
-=======
 [versioneer]
 VCS = git
 style = pep440
@@ -24,4 +5,3 @@ versionfile_source = classo/_version.py
 versionfile_build = classo/_version.py
 tag_prefix =
 parentdir_prefix = c-lasso-
->>>>>>> 8281d58662bb810985f0130817ce6a6edf319463
