+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include versioneer.py
+include classo/_version.py
