+++ .github/workflows/build-and-test.yml
@@ -38,4 +38,5 @@ jobs:
       run: |
         pip install pytest
         pip install pytest-cov
-        pytest tests.py --doctest-modules --junitxml=junit/test-results.xml --cov=com --cov-report=xml --cov-report=html
+        pytest
+        pytest --cov=classo --cov-report=html
