### 2020-11-19 efb7220
Update build-and-test.yml
