+++ classo/solve_R3.py
@@ -3,7 +3,7 @@ import numpy as np
 import numpy.linalg as LA
 from .misc_functions import unpenalized
 
-"""
+r"""
 Problem    :   min ||Ab - y||^2/sigma + n/2 sigma + lambda ||b||1 with C.b= 0 and sigma > 0
 
 Dimensions :   A : m*d  ;  y : m  ;  b : d   ; C : k*d
