+++ .github/workflows/build-and-test.yml
@@ -15,7 +15,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: 3.6
+        python-version: '3.x'
 
     steps:
     - uses: actions/checkout@v2
