+++ .github/workflows/build-and-test.yml
@@ -24,7 +24,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        ls
         pip install -r requirements.txt
     - name: Lint with flake8
       run: |
