### 2020-11-19 leo-simpson/classo#5 @jbytecode
Hi,

There is still uncorrected ones, please check all (for example)

- A space before and after an operator

`(X, C, y), sol = random_data(n, d, d_nonzero, k, sigma, zerosum=True, seed=123)`

becomes

`(X, C, y), sol = random_data(n, d, d_nonzero, k, sigma, zerosum = True, seed = 123)`

and if it is too long, it could be divided into two lines properly.

Note: You can compile and preview the pdf file using [this tool](https://whedon.theoj.org/)
