### 2020-11-19 Leo-Simpson/c-lasso#6 @jbytecode
[JOSS Review]: Performance
Due to paper submission in [JOSS](https://github.com/openjournals/joss-reviews/issues/2844)

There is an item in reviewing list that states

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

Since, ~~the paper does not directly include any performance claims~~, the README.md file includes

> ...c-lasso has a default algorithm setting that proved to be the fastest in our numerical experiments...

If this claim is current, please add a small micro-benchmark script that reports performances of implemented methods so people can reproduce the results and make comparisons.
