### 2020-11-19 Leo-Simpson/c-lasso#7 @jbytecode
[JOSS Review]: Community guidelines
Dear authors,

Due to manuscript submission in [JOSS](https://github.com/openjournals/joss-reviews/issues/2844)

There is an item in reviewing list that states

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

however the repository does not have the addressed items. This [page](https://github.com/Leo-Simpson/c-lasso/community) can be used to create one at least for `newcomers` (1).

Please clearly indicate how to contribute this software, how can users create issues or report errors, and define a contact media (including e-mail or any other one) for those who needs help (I see an e-mail address in documentation, it is okay, but as addressed in the review items, it must be placed in repository in formal ways).

Thank you in advance.
