### 2020-11-19 leo-simpson/classo#6 @jbytecode
Note:

It is also mentioned in paper

> Note that the run time for this d = 100-dimensional example for a single path computation
> is about 0.5 seconds on a standard Laptop.

and this is definitely a performance claim. Please follow the instructions commented above. microbenchmark script is the solution. The standard Laptop is also an ambiguous term. Please specify.
