### 2020-11-19 leo-simpson/classo#6 @muellsen
Thanks for the suggestion! Yes, absolutely. Most of our computational work was actually to test the performance of the different methods. We will add a benchmark and include it, and clarify the run time statements.
