+++ paper/paper.md
@@ -276,7 +276,7 @@ Here, we use a problem instance with $n=100$, $d=100$, a $\beta$ with five non-z
 from classo import classo_problem, random_data
 
 n, d, d_nonzero, k, sigma = 100, 100, 5, 1, 0.5
-(X, C, y), sol = random_data(n, d, d_nonzero, k, sigma, zerosum=True, seed=123)
+(X, C, y), sol = random_data(n, d, d_nonzero, k, sigma, zerosum = True, seed = 123)
 print("Relevant variables  : {}".format(numpy.nonzero(sol)[0]))
 
 problem = classo_problem(X, y, C)
@@ -345,7 +345,7 @@ The code snippet below shows how `c-lasso` is called in `R` to perform regressio
 In `R`, X and C need to be of ```matrix``` type, and y of ```array``` type.
 
 ```r
-problem <- classo$classo_problem(X=X,C=C,y=y) 
+problem <- classo$classo_problem(X = X, C = C, y = y) 
 problem$model_selection$LAMfixed <- TRUE
 problem$model_selection$StabSel <- FALSE
 problem$model_selection$LAMfixedparameters$rescaled_lam <- TRUE
