### 2020-11-22 leo-simpson/classo#5 @Leo-Simpson
Hello,
Those spaces (for = in argument of a function) were not there because those was not taken into account by the package "black" I used to reformat. Now I have haded it in the code and in the examples !

So I think the syntax should be ok this time.
