### 2020-11-22 leo-simpson/classo#5 @jbytecode
seems okay. thank you.
