### 2020-11-22 6299bc5
Update contributing.rst
