+++ docs/source/contributing.rst
@@ -5,14 +5,7 @@ Contributing to c-lasso
 ``c-lasso`` is a package that always can be improved. Any feedback can
 help a lot to fix some bug and to add possible new functionality.
 
-
-Development Portal
-==================
-
-Code development is currently hosted at GitHub.  Issues, feature requests, and
-code contributions are currently handled there.
-
-   https://github.com/Leo-Simpson/c-lasso/
+One can contribute either by reporting  an error, either 
 
 
 Reporting errors
@@ -25,33 +18,15 @@ Any errors or general problems can be reported on GitHub's Issue tracker:
 The quickest way resolve a problem is to go through the following steps:
 
 * Have I tested this on the latest GitHub (``master``) version?
-
-* Have I provided a sample code block which reproduces the error?  Have I
-  tested the code block?
-
-* Have I included the necessary input or output files?
-
-  Sometimes a file attachment is required, since uncommon whitespace or
-  Unicode characters may be missing from a standard cut-and-paste of the file.
-
-* Have I provided a backtrace from the error?
-
-Usually this is enough information to reproduce and resolve the problem.  In
-some cases, we may need more information about your system, including the
-following:
-
-* Your ``classo`` version::
+To see which version you use, you can run on python :
 
      >>> import classo
      >>> classo.__version__
+  
 
-* The version and build of python::
 
-     >>> import sys
-     >>> print(sys.version)
-
-* Your operating system (Linux, OS X, Windows, etc.), and your distribution if
-  relevant.
+* Have I provided a sample code block which reproduces the error?  Have I
+  tested the code block?
 
 While more information can help, the most important step is to report the
 problem, and any missing information can be provided over the course of the
@@ -61,10 +36,7 @@ discussion.
 Feature requests
 ================
 
-Feature requests are welcome, and can be submitted as Issues in the GitHub
-tracker.
-
-   https://github.com/Leo-Simpson/c-lasso/issues
+We recommend opening an issue on `GitHub <https://github.com/Leo-Simpson/c-lasso/issues>`_ to discuss potential changes.
 
 When preparing a feature request, consider providing the following information:
 
@@ -78,54 +50,54 @@ When preparing a feature request, consider providing the following information:
 * Does this feature require any new dependencies ?
 
 
-Contributing to c-lasso
-======================
+
+
+
+
+
+Adding a feature
+==================
+
+One can also contribute with a new feature or with fixing a bug.
+
+Feature should be sent as pull requests via `GitHub <https://github.com/Leo-Simpson/c-lasso>`_, specifically to the
+``master`` branch, which acts as the main development branch.
 
 Fixes and features are very welcome to ``c-lasso``, and are greatly encouraged.
 
 If you are concerned that a project may not be suitable or may conflict with
-ongoing work, then feel free to submit a feature request with comment noting
-that you are happy to provide the feature.
+ongoing work, then feel free to submit a feature request.
 
-Feature should be sent as pull requests via GitHub, specifically to the
-``master`` branch, which acts as the main development branch.
+When preparing a pull request, one should check that the code changes:
 
-Explicit patches via email are also welcome.
+* Pass existing tests, this can be done by running within the root directory:
 
-When preparing a pull request, consider the following advice:
+.. code-block:: bash
 
-* Code changes must pass existing tests::
+  $ pip install --upgrade pytest
+  $ pytest
 
-     $ pytest
-
-* Providing a test case for your example would be greatly appreciated.  See
-  the test files in ``tests``for examples.
+* Includes a test case.
+  See the files in ``c-lasso/tests`` for examples
   
-* Changes must be accompanied by **updated documentation** and examples.
-
-* Features should generally only depend on the standard library.  Any features
[…]
