### 2020-11-22 12d96dd
Update contributing.rst
