### 2020-11-22 d593e4a
Update contributing.rst
