+++ docs/source/contributing.rst
@@ -101,3 +101,9 @@ When preparing a pull request, one should make sure that the code changes:
     $ make html
   
  
+ Contact
+ =======
+ 
+ If the above ways of interacting with c-lasso does not fit your request,
+ you can also `contact directly one of the autors <leo.bill.simpson@gmail.com>`_.
+ 
