### 2020-11-22 b3d4678
Update contributing.rst
