+++ contributing.rst
@@ -104,6 +104,6 @@ When preparing a pull request, one should make sure that the code changes:
 Seeking for support ?
 =======================
 
-If the above ways of interacting with c-lasso does not fit your request,
+If the above ways of interacting with ``c-lasso`` does not fit your request,
 you may `contact directly one of the autors <leo.bill.simpson@gmail.com>`_.
  
