### 2020-11-22 b411a21
Update CONTRIBUTING.rst
