### 2020-11-22 8bef38c
Update CONTRIBUTING.rst
