### 2020-11-22 57e3c21
Update CONTRIBUTING.rst
