+++ CONTRIBUTING.rst
@@ -2,8 +2,6 @@
 Contributing to c-lasso
 ==========================
 
-_`guideline for third parties  <https://github.com/Leo-Simpson/c-lasso/blob/master/CONTRIBUTING.rst>`_
-
 ``c-lasso`` is a package that always can be improved. Any feedback can
 help a lot to fix some bug and to add possible new functionality.
 
