### 2020-11-22 leo-simpson/classo#7 @Leo-Simpson
Hello,

I added a  guideline for third parties  (https://github.com/Leo-Simpson/c-lasso/blob/master/CONTRIBUTING.rst), as well as a basic code of conduct (https://github.com/Leo-Simpson/c-lasso/blob/master/CODE_OF_CONDUCT.md).

I added a contact media in the end of the contributing file. I hope it is what you meant by placing it in formal ways.

Best regards.
