### 2020-11-22 leo-simpson/classo#7 @jbytecode
seems okay, thank you!
