+++ classo/tests/test_compact_func.py
@@ -1,0 +1,7 @@
+import numpy as np
+
+from ..compact_func import (
+    pathlasso,
+    Classo
+)
+
