+++ classo/tests/test_solver.py
@@ -1,0 +1,18 @@
+import numpy as np
+
+from ..solver import (
+    classo_problem,
+    Data,
+    Formulation,
+    Model_selection,
+    PATHparameters,
+    CVparameters,
+    StabSelparameters,
+    LAMfixedparameters,
+    Solution,
+    solution_PATH,
+    solution_CV,
+    solution_StabSel,
+    solution_LAMfixed,
+    choose_numerical_method
+)
