+++ classo/tests/test_misc_func.py
@@ -1,0 +1,26 @@
+import numpy as np
+
+from ..misc_functions import (
+    rescale,
+    theoretical_lam,
+    min_LS,
+    affichage,
+    check_size,
+    tree_to_matrix,
+    random_data,
+    csv_to_np,
+    mat_to_np,
+    clr,
+    to_zarr,
+    unpenalized,
+    plot_betai,
+    influence,
+    normalize,
+    denorm,
+    hub,
+    L_LS,
+    L_conc,
+    L_H,
+    proj_c,
+    remove_same_vect
+)
