+++ classo/tests/test_path_alg.py
@@ -1,0 +1,23 @@
+import numpy as np
+
+from ..path_alg import (
+    parameters_for_update,
+    solve_path,
+    solve_path_Conc,
+    pathalgo_general,
+    pathalgo_huber_cl,
+    pathalgo_cl,
+    up_LS,
+    up_huber,
+    up_cl,
+    up_huber_cl,
+    derivatives,
+    next_idr1,
+    next_idr2,
+    h_lambdamax,
+    h_prime,
+    find_F,
+    find_beta0,
+    binary_search
+)
+
