+++ classo/tests/test_solve_R1.py
@@ -1,0 +1,10 @@
+import numpy as np
+
+from ..solve_R1 import (
+    prox,
+    proj_c,
+    QQ,
+    pathlasso_R1,
+    Classo_R1,
+    problem_R1
+)

[…]
