### 2020-11-23 a4a7f8b
Update cross_validation.py
