### 2020-11-23 da168d8
Update cross_validation.py
