### 2020-11-30 leo-simpson/classo#4 @jbytecode
Dear authors,

Since there is not any progress for this issue, I am writing here for just checking. Please let me know if you are in trouble.

Have a nice week.
