### 2020-11-30 leo-simpson/classo#4 @Leo-Simpson
Dear @jbytecode,

I am still working on it, I will push my test files when they are ready, by the end of the week I hope, if it is ok.

Best regards.
