+++ .github/workflows/build-and-test.yml
@@ -35,5 +35,4 @@ jobs:
       run: |
         pip install pytest
         pip install pytest-cov
-        pytest
-        pytest --cov-report=html
+        py.test --cov --cov-report html --cov-config=.coveragerc
