### 2020-12-03 leo-simpson/classo#4 @Leo-Simpson
Dear @jbytecode ,

Automated tests for all functionalities are ready. There is also the GitHub action associated, so that the tests are automatically done at each commit.

Best regards.
