### 2020-12-03 leo-simpson/classo#4 @jbytecode
Thank you.

One more issue: I couldn't see any tests performed in file test_solver.py. Just function calls to solve(). Isnt it possible to check some values after performing this function?
