### 2020-12-03 leo-simpson/classo#4 @jbytecode
One more issue: are you sure the tests are triggered in each single commit? What does ensure that in the action file? Yes it is right there is an action file but I couldnt see any tests performed. Just asking for checking.
