### 2020-12-03 leo-simpson/classo#4 @jbytecode
Okay, I now see the stage

```
classo/tests/test_Classo_R.py ....                                       [  7%]
classo/tests/test_compact_func.py ..........                             [ 26%]
classo/tests/test_cross_validation.py ......                             [ 37%]
classo/tests/test_misc_func.py ..........                                [ 56%]
classo/tests/test_solver.py ............                                 [ 79%]
classo/tests/test_stability_selection.py ...........                     [100%]
```

I am closing this issue, but please add tests in test_solver.py if there is not any.
