### 2020-12-03 leo-simpson/classo#6 @jbytecode
Dear authors,

This is the last issue. Whenever you finish this work, I will finalize my review.

Thank you for your great effort.
