### 2020-12-03 leo-simpson/classo#4 @Leo-Simpson
Thank you for closing the issue.

I just wanted to mention that the correctness of the algorithms are checked more in test_Classo_R.py , because that is where it is checked that the different optimization scheme gives the same result (or a close result at least).

I can also add some checks in in the test of solver, but the idea there was just to check that everything assembled with no mistakes.

Also, just wanted to add that I will try to automatically add the html coverage report that is generated by the action to the repo, because this informs the percentage of the code that is covered.

Best regards.
