### 2020-12-09 #2844 @glemaitre
I am adding a question here since that I would like to have advice from @mjsottile

So one statement in the package is the following:

> his package is intended to fill the gap between popular python tools such as scikit-learn which CANNOT solve sparse constrained problems and general-purpose optimization solvers that do not scale well for the considered problems.

 It is completely correct that scikit-learn does not provide this functionality. So, one of my expectation would be to be able to use `c-lasso` as a drop-in replacement instead of scikit-learn to solve such problem. However, I might be bias by being involve in the development of scikit-learn. The issue here is that you cannot use the package for such purpose due to API issue incompatibilities.

Indeed, it would require a tremendous amount of work but this is possible. I can link to the following package which shows such of an example (https://mathurinm.github.io/celer/index.html). I would expect Python functions to implement the different solvers and a common class with parameters allowing to select a specific solver. Turning on/off flags is not really something done in scikit-learn. With such design, `c-lasso` would be used by people from the scikit-learn community while currently it would not be possible.

I see that @Leo-Simpson is also using the package with R which is probably one of the reason of having different vison regarding the package.

@mjsottile Indeed, I would like to know if these thoughts should be considered while reviewing the paper. I would not like to block the review/acceptance process on a criterion that should not be considered in the review process.
