### 2020-12-09 leo-simpson/classo#6 @muellsen
Update regarding performance benchmarks: We will provide this in the upcoming days, including a comparison to cvx. Will ping you once done!
