### 2020-12-09 #2844 @muellsen
Thanks for raising this important issue. Indeed, our initial hope was to design c-lasso as drop-in replacement in scikit-learn and use all the cool functionality there. However, after an in-depth analysis at the start of the project in early 2019, we opted for a stand-alone version. Below is our thinking and a note about potential future developments
- First, we do not have yet the resources and knowledge to do such an integration. I spent several days understanding the implementations of the optimization schemes underlying the standard lasso problems and found them to be incompatible with our ideas (largely due to my inadequacies I think)
- Second, c-lasso is indeed integral for several R packages (including one already available, called "trac") that we are currently developing, and thus, the current solution makes it easier (for us) to handle these joint developments
- Third, c-lasso is integral for future QIIME-2  plugins (developed in parallel) where we again would love to have light-weight dependencies and full control over the solver development (this is specific to the needs in the microbiome data analysis community)
- Forth, if someone from the scikit-learn community is interested in this, we would love to collaborate. I know that Alexandre Gramfort and colleagues have worked on similar ideas regarding concomitant estimation (so problems with estimating sigma) so maybe there is a future connection?!
