### 2020-12-09 #2844 @mjsottile
Hi @glemaitre - good questions, and thanks for the details @muellsen.  I think this issue is worth addressing in the paper explicitly clarify that, in its current state, the package is not an API-compatible drop in replacement for components of scikit-learn.  The issue with API compatibility (or the lack thereof) doesn't detract from the quality of the work itself, just the expectations that users may have.  If the paper just adds a simple statement that the code currently is not a straight drop-in replacement in scikit-learn, then potential users won't mistakenly assume that it is.  I think such a design for c-lasso would be interesting to look at in the future, but isn't critical for this submission.
