### 2020-12-09 #2844 @glemaitre
OK, so this is fine with me. I just open a couple of issues such that the content of the README is included in the documentation webpage and to have a second look regarding the test coverage.

Now I am hooked so I am might propose a bit of help regarding the scikit-learn API part :)
