### 2020-12-10 b953f41
add  readme info into doc; rerun examples; change to relative path
