+++ docs/source/getting-started.rst
@@ -1,0 +1,415 @@
+
+Getting started
+======================
+
+
+
+
+Installation
+^^^^^^^^^^^^^^^^
+
+c-lasso is available on pip. You can install the package
+in the shell using
+
+.. code-block::
+
+     pip install c-lasso
+
+To use the c-lasso package in Python, type 
+
+.. code-block:: python
+
+     from classo import classo_problem 
+     # one can add auxiliary functions as well such as random_data or csv_to_np
+
+
+The `c-lasso` package depends on the following Python packages:
+
+- ``numpy``; 
+- ``matplotlib``; 
+- ``scipy``; 
+- ``pandas``; 
+- ``h5py``
+- ``pytest`` (for tests)
+
+
+
+Examples on synthetic data
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+
+Basic example
+""""""""""""""""
+
+
+We begin with a basic example that shows how to run c-lasso on synthetic data. This example and the next one can be found on the notebook 'Synthetic data Notebook.ipynb'
+
+The c-lasso package includes
+the routine ``random_data`` that allows you to generate problem instances using normally distributed data.
+
+.. code-block:: python
+
+     m, d, d_nonzero, k, sigma = 100, 200, 5, 1, 0.5
+     (X, C, y), sol = random_data(m, d, d_nonzero, k, sigma, zerosum=True, seed=1)
+
+This code snippet generates a problem instance with sparse :math:`\beta` in dimension
+d=100 (sparsity d_nonzero=5). The design matrix X comprises n=100 samples generated from an i.i.d standard normal
+distribution. The dimension of the constraint matrix C is d x k matrix. The noise level is :math:`\sigma=0.5`. 
+The input ``zerosum=True`` implies that C is the all-ones vector and :math:`C\beta=0`. The n-dimensional outcome vector y
+and the regression vector :math:`\beta` is then generated to satisfy the given constraints. 
+
+Next we can define a default c-lasso problem instance with the generated data:
+
+.. code-block:: python
+
+     problem = classo_problem(X, y, C) 
+
+You can look at the generated problem instance by typing:
+
+.. code-block:: python
+
+     print(problem)
+
+
+This gives you a summary of the form:
+
+.. code-block::
+
+     FORMULATION: R3
+     
+     MODEL SELECTION COMPUTED:  
+          Stability selection
+     
+     STABILITY SELECTION PARAMETERS: 
+          numerical_method : not specified
+          method : first
+          B = 50
+          q = 10
+          percent_nS = 0.5
+          threshold = 0.7
+          lamin = 0.01
+          Nlam = 50
+
+As we have not specified any problem, algorithm, or model selection settings, this problem instance
+represents the *default* settings for a c-lasso instance: 
+- The problem is of regression type and uses formulation [R3], i.e. with concomitant scale estimation. 
+- The *default* optimization scheme is the path algorithm (see [Optimization schemes](#optimization-schemes) for further info). 
+- For model selection, stability selection at a theoretically derived :math:`\beta` value is used (see [Reference](#references) [4] for details). Stability selection comprises a relatively large number of parameters. For a description of the settings, we refer to the more advanced examples below and the API.
+
+You can solve the corresponding c-lasso problem instance using
+
+.. code-block:: python
+
+     problem.solve()
+
+
+After completion, the results of the optimization and model selection routines 
+can be visualized using
+
[…]
