+++ docs/source/description.rst
@@ -1,0 +1,158 @@
+
+Mathematical description
+=============================
+
+
+The forward model is assumed to be: 
+
+.. math::
+   y = X \beta + \sigma \epsilon \qquad \textrm{subject to} \qquad C\beta=0
+
+Here, y and X are given outcome and predictor data. The vector y can be continuous (for regression) or binary (for classification). C is a general constraint matrix. The vector :math:`\beta` comprises the unknown coefficients and :math:`\sigma` an 
+unknown scale.
+
+The package handles several different estimators for inferring :math:`\beta` and :math:`\sigma`), including 
+the constrained Lasso, the constrained scaled Lasso, and sparse Huber M-estimation with linear equality constraints.
+Several different algorithmic strategies, including path and proximal splitting algorithms, are implemented to solve 
+the underlying convex optimization problems.
+
+We also include two model selection strategies for determining the sparsity of the model parameters: k-fold cross-validation and stability selection.   
+
+This package is intended to fill the gap between popular python tools such as `scikit-learn <https://scikit-learn.org/stable/>`_ which CANNOT solve sparse constrained problems and general-purpose optimization solvers that do not scale well for the considered problems.
+
+Below we show several use cases of the package, including an application of sparse *log-contrast*
+regression tasks for *compositional* microbiome data.
+
+The code builds on results from several papers which can be found in the [References](#references). We also refer to the accompanying `JOSS paper submission <https://github.com/Leo-Simpson/c-lasso/blob/master/paper/paper.md>`_, also available on `arXiv <https://arxiv.org/pdf/2011.00898.pdf>`_.
+
+
+Regression and classification problems
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+The c-lasso package can solve six different types of estimation problems: 
+four regression-type and two classification-type formulations.
+
+[R1] Standard constrained Lasso regression
+"""""""""""""""""""""""""""""""""""""""""""""          
+
+.. math::
+   \min_{\beta \in \mathbb{R}^d} \left\lVert X\beta - y \right\rVert^2 + \lambda \left\lVert \beta\right\rVert_1 \qquad \textrm{subject to} \qquad  C\beta = 0
+
+
+This is the standard Lasso problem with linear equality constraints on the :math:`\beta` vector. 
+The objective function combines Least-Squares for model fitting with l1 penalty for sparsity.   
+
+[R2] Contrained sparse Huber regression
+""""""""""""""""""""""""""""""""""""""""""""""""""                   
+
+.. math::
+   \min_{\beta \in \mathbb{R}^d} h_{\rho} (X\beta - y) + \lambda \left\lVert \beta\right\rVert_1 \qquad \textrm{subject to} \qquad  C\beta = 0
+
+This regression problem uses the `Huber loss <https://en.wikipedia.org/wiki/Huber_loss>`_ as objective function 
+for robust model fitting with l1 and linear equality constraints on the :math:`\beta` vector. The parameter :math:`\rho=1.345`.
+
+[R3] Contrained scaled Lasso regression
+""""""""""""""""""""""""""""""""""""""""""""""""""  
+
+.. math::
[…]
