+++ docs/source/index.rst
@@ -17,11 +17,14 @@ The package is available on `https://github.com/Leo-Simpson/c-lasso`.
 
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 2
    :caption: Contents:
 
+   description
+   getting-started
    problem_classo
    misc-functions
+   structure-code
 
    additional/license
    additional/help
