+++ docs/source/structure-code.rst
@@ -1,0 +1,7 @@
+Structure of the code
+========================
+
+
+.. figure:: additional/classo_structure.png
+   :width: 600 px
+   :align: center

[…]
