### 2020-12-10 ac91e8b
move contributing.rst to the documentation
