+++ docs/source/additional/help.rst
@@ -1,9 +1,0 @@
-Seeking support ?
-==================
-
-For more information about the ``c-lasso``, please check the `GitHub <https://github.com/Leo-Simpson/c-lasso>`_ webpage.
-
-If you want to report a problem or propose a suggestion, please open an issue `issue <https://github.com/Leo-Simpson/c-lasso/issues>`_.
-
-If the above ways of interacting with ``c-lasso`` does not fit your request,
-you may `contact directly one of the autors <leo.bill.simpson@gmail.com>`_.
