+++ examples/README.txt
@@ -1,0 +1,5 @@
+
+Examples Gallery
+================
+
+Bellow is a gallery of examples.
