### 2020-12-11 8129b48
update documented examples
