+++ setup.py
@@ -12,12 +12,34 @@ setup(name='c-lasso',
       packages=['classo'],
       install_requires = [
                           'numpy',
-                          'matplotlib',
-                          'pandas',
                           'h5py',
-                          'scipy',
-                          'pytest'
+                          'scipy'
                     ],
+      extras_require = {
+            'tests': [
+                  'pytest',
+                  'pytest-cov'],
+            'docs': [
+                  'sphinx',
+                  'sphinx-gallery',
+                  'sphinx_rtd_theme',
+                  'sphinxcontrib-bibtex',
+                  'numpydoc',
+                  'matplotlib',
+                  'pandas',
+            ]
+      },
+      classifiers=['Intended Audience :: Science/Research',
+                  'Operating System :: Microsoft :: Windows',
+                  'Operating System :: POSIX',
+                  'Operating System :: Unix',
+                  'Operating System :: MacOS',
+                  'Programming Language :: Python :: 3',
+                  'Programming Language :: Python :: 3.6',
+                  'Programming Language :: Python :: 3.7',
+                  'Programming Language :: Python :: 3.8',
+                  'Programming Language :: Python :: 3.9',
+                  ],
       long_description_content_type = 'text/markdown',
       long_description=open('README.md').read(),
       zip_safe=False)
