+++ classo/solve_R3.py
@@ -152,7 +152,7 @@ def pathlasso_R3(pb, path, n_active = False):
     if type(n_active) == int and n_active > 0:
         n_act = n_active
     else:
-        n_act = n
+        n_act = d
 
     for lam in path:
         X = Classo_R3(pb, lam)
