+++ classo/solve_R4.py
@@ -140,7 +140,7 @@ def pathlasso_R4(pb, path, n_active = False):
     if type(n_active) == int and n_active > 0:
         n_act = n_active
     else:
-        n_act = d
+        n_act = d + 1
 
     for lam in path:
         X = Classo_R4(pb, lam)
