+++ classo/solve_R2.py
@@ -197,7 +197,7 @@ def pathlasso_R2(pb, path, n_active = False):
     if type(n_active) == int and n_active > 0:
         n_act = n_active
     else:
-        n_act = d
+        n_act = d + 1
     for lam in path:
         X = Classo_R2(pb, lam, compute = False)
         BETA.append(X[0])
