+++ examples/plot_basic_example.py
@@ -3,6 +3,7 @@ Basic example
 ===============
 
 Let's present what classo does when using its default parameters on synthetic data.
+
 """
 
 from classo import classo_problem, random_data
@@ -11,6 +12,7 @@ import numpy as np
 # %%
 #  Generate the data
 # ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+# 
 # This code snippet generates a problem instance with sparse ß in dimension
 # d=100 (sparsity d_nonzero=5). The design matrix X comprises n=100 samples generated from an i.i.d standard normal
 # distribution. The dimension of the constraint matrix C is d x k matrix. The noise level is σ=0.5. 
@@ -22,11 +24,13 @@ m, d, d_nonzero, k, sigma = 100, 200, 5, 1, 0.5
 
 # %%
 # Remark : one can see the parameters that should be selected :
+
 print(np.nonzero(sol))
 
 # %%
 # Define the classo instance
 # ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+#  
 # Next we can define a default c-lasso problem instance with the generated data:
 
 problem = classo_problem(X, y, C) 
@@ -34,6 +38,7 @@ problem = classo_problem(X, y, C)
 # %%
 # Check parameters
 # ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+#  
 # You can look at the generated problem instance by typing:
 
 print(problem)
@@ -41,7 +46,8 @@ print(problem)
 # %%
 #  Solve optimization problems
 # ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-#  We only use stability selection as default model selection strategy. 
+#  
+# We only use stability selection as default model selection strategy. 
 # The command also allows you to inspect the computed stability profile for all variables 
 # at the theoretical λ
 
@@ -50,6 +56,7 @@ problem.solve()
 # %%
 # Visualisation
 # ^^^^^^^^^^^^^^^
+#  
 # After completion, the results of the optimization and model selection routines 
 # can be visualized using
 
