+++ docs/source/conf.py
@@ -18,7 +18,6 @@ import sphinx_gallery
 import sys
 import warnings
 
-warnings.filterwarnings("ignore")
 
 # -- Project information -----------------------------------------------------
 
@@ -61,7 +60,7 @@ templates_path = ['_templates']
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ['_build', '.DS_Store']
 
 
 # -- Options for HTML output -------------------------------------------------
