+++ requirements.txt
@@ -8,3 +8,4 @@ pytest
 pytest-cov
 sphinx
 sphinx-gallery
+sphinx_rtd_theme
