+++ docs/source/auto_examples/sg_execution_times.rst
@@ -1,18 +1,0 @@
-
-:orphan:
-
-.. _sphx_glr_auto_examples_sg_execution_times:
-
-Computation times
-=================
-**00:08.104** total execution time for **auto_examples** files:
-
-+---------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_plot_advanced_example.py` (``plot_advanced_example.py``) | 00:08.104 | 0.0 MB |
-+---------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_plot_basic_example.py` (``plot_basic_example.py``)       | 00:00.000 | 0.0 MB |
-+---------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_plot_combo_example.py` (``plot_combo_example.py``)       | 00:00.000 | 0.0 MB |
-+---------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_examples_plot_pH_example.py` (``plot_pH_example.py``)             | 00:00.000 | 0.0 MB |
-+---------------------------------------------------------------------------------------+-----------+--------+

[…]
