+++ README.md
@@ -381,15 +381,13 @@ print(problem.solution)
 The next microbiome example considers the [Central Park Soil dataset](./examples/pH_data) from [Ramirez et al.](https://royalsocietypublishing.org/doi/full/10.1098/rspb.2014.1988). The sample locations are shown in the Figure on the right.)
 -->
 
-#### pH prediction using soil dataset
+#### pH prediction using the 88 soils dataset
 
-The next microbiome example considers a [Soil dataset](./examples/pH_data).
+The next microbiome example considers the [88 soils dataset](./examples/pH_data) from [Lauber et al., 2009](https://pubmed.ncbi.nlm.nih.gov/19502440/).
 
-The task is to predict pH concentration in the soil from microbial abundance data. 
-
-This task was also considered in [Tree-Aggregated Predictive Modeling of Microbiome Data](https://www.biorxiv.org/content/10.1101/2020.09.01.277632v1) for [another sol dataset](https://royalsocietypublishing.org/doi/full/10.1098/rspb.2014.1988).
-
-
+The task is to predict pH concentration in the soil from microbial abundance data. A similar analysis is available
+in [Tree-Aggregated Predictive Modeling of Microbiome Data](https://www.biorxiv.org/content/10.1101/2020.09.01.277632v1) 
+with Central Park soil data from [Ramirez et al.](https://royalsocietypublishing.org/doi/full/10.1098/rspb.2014.1988).
 
 Code to run this application is available in [the accompanying notebook](./examples/example-notebook.ipynb) under `pH data`. Below is a summary of a c-lasso problem instance (using the R3 formulation).
  
