+++ examples/README.txt
@@ -2,4 +2,4 @@
 Examples Gallery
 ================
 
-Bellow is a gallery of examples.
+Below is a gallery of examples.
