### 2020-12-14 807942a
Update .readthedocs.yml
