+++ .readthedocs.yml
@@ -22,3 +22,4 @@ python:
   version: 3.8
   install:
     - requirements: requirements.txt
+    - requirements: c-lasso
