### 2020-12-14 e96657c
Update .readthedocs.yml
