+++ examples/example_PH.py
@@ -4,8 +4,8 @@ import numpy as np
 from copy import deepcopy as dc
 import scipy.io as sio
 
-pH = sio.loadmat("pH_data/pHData.mat")
-tax = sio.loadmat("pH_data/taxTablepHData.mat")["None"][0]
+pH = sio.loadmat("pH_data/matlab/pHData.mat")
+tax = sio.loadmat("pH_data/matlab/taxTablepHData.mat")["None"][0]
 
 X, Y_uncent = pH["X"], pH["Y"].T[0]
 y = Y_uncent - np.mean(Y_uncent)  # Center Y
