+++ docs/source/auto_examples/plot_Tara_example.py
@@ -86,6 +86,7 @@ X = np.log(pseudo_count+x)
 nleaves = np.sum(A,axis = 0)
 logGeom = X.dot(A)/nleaves
 
+raise ValueError("Do not compute it now")
 # %%
 # Cross validation and Path Computation
 # ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
