+++ examples/pH_data/qiime2/originals/88soils_modified_metadata.txt
@@ -1,0 +1,90 @@
+#SampleID	BarcodeSequence	LinkerPrimerSequence	barcode_read_group_tag	dna_extracted_prep	experiment_alias	experiment_center	experiment_design_description	experiment_title	instrument_name	key_seq	library_construction_protocol	linker	pcr_primers	physical_specimen_remaining_prep	platform	pool_member_name	pool_proportion	primer_read_group_tag	region	run_alias	run_center	run_date	run_prefix	samp_size	sample_center	sample_type_prep	sequencing_meth	study_center	study_ref	target_gene	target_subfragment	altitude	annual_season_precpt	annual_season_temp	anonymized_name	assigned_from_geo	carb_nitro_ratio	cmin_rate	collection_date	common_name	country	depth	dna_extracted	elevation	env_biome	env_feature	env_matter	host_subject_id	latitude	longitude	ph	physical_specimen_remaining	project_name	public	sample_type	silt_clay	soil_moisture_deficit	soil_type	specific_location	taxon_id	texture	title	tot_org_carb	tot_org_nitro	Description	ph2	ph3	ph4	ph_rounded
+103.BZ1	ACACATGTCTAC	CATGCTGCCTCCCGTAGGAGT	BZ1_V2	True	lauber_88_soils	CCME	bacterial biogeography	Pyrosequencing_based_assessment_of_soil_pH_as_a_predictor_of_soil_bacterial_community_structure_at_the_continental_scale	Engencore	TCAG	19502440	CA	FWD:GCCTTGCCAGCCCGCTCAGTCAGAGTTTGATCCTGGCTCAG;REV:CATGCTGCCTCCCGTAGGAGT	False	FLX	BZ1_V2	1	V2	0	BZ1_V2	CCME	5/28/08	FA6P1OK	0.5, g	CCME	XXQIITAXX	pyrosequencing	CCME	lauber_88_soils	16S rRNA	V2	0	260	-2.9	BZ1	False	19.4	4.185	2008	soil metagenome	GAZ:United States of America	0-0.05	True	300	ENVO:forest	ENVO:forest soil	ENVO:soil	BZ1	64.8	-148.25	5.12	False	lauber_88_soils	True	XXQIITAXX	80	132.6	gelisol	Bonanza Creek LTER, AK USA	410658	silt loam	Pyrosequencing-Based Assessment of Soil pH as a Predictor of Soil Bacterial Community Structure at the Continental Scale	30.3	0.0	soil metagenome	26.2144	134.21772800000002	687.19476736	5_
[…]
