+++ docs/source/auto_examples/plot_pH_example.rst
@@ -25,15 +25,10 @@ A similar analysis is also done in `Tree-Aggregated Predictive Modeling of Micro
 
 
     from classo import classo_problem
+
+    import qiime2
+    import pandas as pd
     import numpy as np
-    from copy import deepcopy as dc
-    import scipy.io as sio
-
-
-
-
-
-
 
 
 Load data
@@ -43,26 +38,17 @@ Load data
 .. code-block:: default
 
 
-    pH = sio.loadmat("pH_data/pHData.mat")
-    tax = sio.loadmat("pH_data/taxTablepHData.mat")["None"][0]
+    table_art = qiime2.Artifact.load('pH_data/news/88soils_balances.qza')
+    table = table_art.view(pd.DataFrame)
+    metadata = pd.read_table('pH_data/originals/88soils_modified_metadata.txt', index_col=0)
+    label = list(table.index)
 
-    X, Y_uncent = pH["X"], pH["Y"].T[0]
+
+    X = np.array(table)
+    y_uncent = metadata["ph"].values
     y = Y_uncent - np.mean(Y_uncent)  # Center Y
     print(X.shape)
-
-
-
-
-
-.. rst-class:: sphx-glr-script-out
-
- Out:
-
- .. code-block:: none
-
-    (88, 116)
-
-
+    print(y.shape)
 
 
 Set up c-lassso problem
@@ -72,18 +58,12 @@ Set up c-lassso problem
 .. code-block:: default
 
 
-    problem = classo_problem(X, y) 
+    problem = classo_problem(X, y, label = label) 
 
     problem.model_selection.StabSelparameters.method      = 'lam'
     problem.model_selection.PATH = True
     problem.model_selection.LAMfixed = True
     problem.model_selection.PATHparameters.n_active = X.shape[1] + 1
-
-
-
-
-
-
 
 
 Solve for R1
@@ -95,90 +75,6 @@ Solve for R1
     problem.formulation.concomitant = False
     problem.solve()
     print(problem, problem.solution)
-
-
-
-
-.. rst-class:: sphx-glr-horizontal
-
-
-    *
-
-      .. image:: /auto_examples/images/sphx_glr_plot_pH_example_001.png
-          :alt: Coefficients at $\lambda$ = 0.218
-          :class: sphx-glr-multi-img
-
-    *
-
-      .. image:: /auto_examples/images/sphx_glr_plot_pH_example_002.png
-          :alt: Coefficients across $\lambda$-path using R1
-          :class: sphx-glr-multi-img
-
-    *
-
-      .. image:: /auto_examples/images/sphx_glr_plot_pH_example_003.png
-          :alt: Stability selection profile of type lam using R1
-          :class: sphx-glr-multi-img
-
-    *
-
-      .. image:: /auto_examples/images/sphx_glr_plot_pH_example_004.png
-          :alt: Refitted coefficients after stability selection
-          :class: sphx-glr-multi-img
-
-
-.. rst-class:: sphx-glr-script-out
-
- Out:
-
- .. code-block:: none
-
- 
- 
-    FORMULATION: R1
- 
-    MODEL SELECTION COMPUTED:  
-         Lambda fixed
-         Path
-         Stability selection
- 
-    LAMBDA FIXED PARAMETERS: 
-         numerical_method = Path-Alg
-         rescaled lam : True
-         threshold = 0.008
-         lam : theoretical
-         theoretical_lam = 0.2182
- 
-    PATH PARAMETERS: 
-         numerical_method : Path-Alg
-         lamin = 0.001
-         Nlam = 80
- 
-         maximum active variables = 117
- 
-    STABILITY SELECTION PARAMETERS: 
-         numerical_method : Path-Alg
-         method : lam
-         B = 50
-         q = 10
-         percent_nS = 0.5
[…]

[…]
