### 2020-12-14 a39367b
Update .readthedocs.yml
