+++ .readthedocs.yml
@@ -27,4 +27,4 @@ python:
     - method: pip
       path: .
       extra_requirements:
-        - r2py
+        - rpy2
