+++ .readthedocs.yml
@@ -22,8 +22,6 @@ python:
   version: 3.8
   install:
     - requirements: requirements.txt
+    - requirements: docs/requirements.txt
     - method: setuptools
       path: .
-    - method: pip
-      extra_requirements:
-        - rpy2
