### 2020-12-14 7e35625
Update plot_pH_example.py
