+++ .readthedocs.yml
@@ -22,6 +22,5 @@ python:
   version: 3.8
   install:
     - requirements: requirements.txt
-    - requirements: docs/requirements.txt
     - method: setuptools
       path: .
