### 2020-12-14 7f1ad27
Update plot_Tara_example.py
