### 2020-12-14 5d9fcc6
Update plot_Tara_example.py
