### 2020-12-14 12c919c
Update plot_Tara_example.py
