### 2020-12-15 dfc00cb
first benchmark and add examples
