+++ docs/source/auto_examples/index.rst
@@ -56,27 +56,6 @@ Below is a gallery of examples.
 
 .. raw:: html
 
-    <div class="sphx-glr-thumbcontainer" tooltip="The next microbiome example considers a `Soil dataset &lt;https://github.com/Leo-Simpson/c-lasso/t...">
-
-.. only:: html
-
- .. figure:: /auto_examples/images/thumb/sphx_glr_plot_pH_example_thumb.png
-     :alt: pH prediction using the 88 soils dataset
-
-     :ref:`sphx_glr_auto_examples_plot_pH_example.py`
-
-.. raw:: html
-
-    </div>
-
-
-.. toctree::
-   :hidden:
-
-   /auto_examples/plot_pH_example
-
-.. raw:: html
-
     <div class="sphx-glr-thumbcontainer" tooltip="We first consider the `COMBO data set &lt;https://github.com/Leo-Simpson/c-lasso/tree/master/examp...">
 
 .. only:: html
@@ -95,6 +74,27 @@ Below is a gallery of examples.
    :hidden:
 
    /auto_examples/plot_combo_example
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="The next microbiome example considers a `Soil dataset &lt;https://github.com/Leo-Simpson/c-lasso/t...">
+
+.. only:: html
+
+ .. figure:: /auto_examples/images/thumb/sphx_glr_plot_pH_example_thumb.png
+     :alt: pH prediction using the 88 soils dataset
+
+     :ref:`sphx_glr_auto_examples_plot_pH_example.py`
+
+.. raw:: html
+
+    </div>
+
+
+.. toctree::
+   :hidden:
+
+   /auto_examples/plot_pH_example
 
 .. raw:: html
 

[…]
