+++ benchmark/R1/bm-R1-plot.py
@@ -15,8 +15,8 @@ for name in ["T_pa", "T_pds", "T_dr", "T_cvx"]:
     N_size, N_data = T.shape
     mean = np.mean(T, axis=1)
     std = np.std(T, axis=1)/np.sqrt(N_data)
-
-    plt.errorbar(range(N_size), mean, yerr=std, label=name)
+    label = name[2:]
+    plt.errorbar(range(N_size), mean, yerr=std, label=label)
 
 plt.xticks(range(N_size), labels)
 plt.legend()
@@ -24,15 +24,29 @@ plt.savefig("bm-R1-times.png")
 plt.show()
 
 
-for name in ["L_pa", "L_pds", "L_dr", "L_cvs"]:
+for name in ["L_pa", "L_pds", "L_dr", "L_cvx"]:
     L = data[name] - data["L_pa"]
     N_size, N_data = L.shape
     mean = np.mean(L, axis=1)
     std = np.std(L, axis=1)/np.sqrt(N_data)
-
-    plt.errorbar(range(N_size), mean, yerr=std, label=name)
+    label = name[2:]
+    plt.errorbar(range(N_size), mean, yerr=std, label=label)
 
 plt.xticks(range(N_size), labels)
 plt.legend()
 plt.savefig("bm-R1-losses.png")
+plt.show()
+
+for name in ["C_pa", "C_pds", "C_dr", "Cs_cvx"]:
+    L = data[name] - data["L_pa"]
+    N_size, N_data = L.shape
+    mean = np.mean(L, axis=1)
+    std = np.std(L, axis=1)/np.sqrt(N_data)
+    label = name[2:]
+    plt.errorbar(range(N_size), mean, yerr=std, label=label)
+
+plt.title("Value of norm(Cbeta)")
+plt.xticks(range(N_size), labels)
+plt.legend()
+plt.savefig("bm-R1-constraint.png")
 plt.show()

[…]
