+++ benchmark/C1/bm-C1.py
@@ -1,136 +1,0 @@
-import numpy as np
-import numpy.linalg as LA
-from classo import classo_problem, random_data
-import cvxpy as cp
-from time import time
-
-l = [1, 2, 5, 7]
-
-
-
-def loss(X, y, lam, beta):
-    lamb = lam*2*LA.norm(X.T.dot(y),np.infty)
-    return np.sum((X.dot(beta) - y)**2) + lamb*np.sum(abs(beta))
-
-
-d_nonzero = 5
-sigma = 0.5
-lam = 0.1
-
-N_per_data = 5
-N_data = 10
-SIZES = [
-    (50, 100),
-    (100, 100),
-    (100, 200),
-    (200, 200)
-]
-
-N_sizes = len(SIZES)
-
-T_pa = np.zeros((N_sizes, N_data))
-L_pa = np.zeros((N_sizes, N_data))
-
-T_pds = np.zeros((N_sizes, N_data))
-L_pds = np.zeros((N_sizes, N_data))
-
-T_dr = np.zeros((N_sizes, N_data))
-L_dr = np.zeros((N_sizes, N_data))
-
-T_cvx = np.zeros((N_sizes, N_data))
-L_cvx = np.zeros((N_sizes, N_data))
-
-
-
-for s in range(N_sizes):
-
-    m, d = SIZES[s]
-
-
-    for i in range(N_data):
-        (X, C, y), sol = random_data(m, d, d_nonzero, 1, sigma, zerosum=True, seed=i)
-        lamb = lam*2*LA.norm(X.T.dot(y),np.infty)
-
-        t0 = time()
-        # classo Path-Alg
-        b_pa = []
-        for j in range(N_per_data):
-            problem = classo_problem(X, y, C)
-            problem.formulation.concomitant = False
-            problem.model_selection.StabSel = False
-            problem.model_selection.LAMfixed = True
-            problem.model_selection.LAMfixedparameters.rescaled_lam = True
-            problem.model_selection.LAMfixedparameters.lam = lam
-            problem.model_selection.LAMfixedparameters.numerical_method = 'Path-Alg'
-            problem.solve()
-            b_pa.append(problem.solution.LAMfixed.beta)
-        b_pa = np.array(b_pa)
-
-        t1 = time()
-        # classo P-PDS
-        b_pds = []
-        for j in range(N_per_data):
-            problem = classo_problem(X, y, C)
-            problem.formulation.concomitant = False
-            problem.model_selection.StabSel = False
-            problem.model_selection.LAMfixed = True
-            problem.model_selection.LAMfixedparameters.rescaled_lam = True
-            problem.model_selection.LAMfixedparameters.lam = lam
-            problem.model_selection.LAMfixedparameters.numerical_method = 'P-PDS'
-            problem.solve()
-            b_pds.append(problem.solution.LAMfixed.beta)
-        b_pds = np.array(b_pds)
-
-        t2 = time()
-        # classo DR
-        b_dr = []
-        for j in range(N_per_data):
-            problem = classo_problem(X, y, C)
-            problem.formulation.concomitant = False
-            problem.model_selection.StabSel = False
-            problem.model_selection.LAMfixed = True
-            problem.model_selection.LAMfixedparameters.rescaled_lam = True
-            problem.model_selection.LAMfixedparameters.lam = lam
-            problem.model_selection.LAMfixedparameters.numerical_method = 'P-PDS'
-            problem.solve()
[…]
