### 2020-12-15 c764ca2
Update bm-R2.py
