+++ benchmark/Makefile
@@ -1,0 +1,9 @@
+# Minimal makefile for benchmark
+#
+
+python3.9 R1/bm-R1.py
+python3.9 R1/bm-R1-plot.py
+
+python3.9 R2/bm-R2.py
+python3.9 R2/bm-R2-plot.py
+
