+++ benchmark/R2/bm-R2-plot.py
@@ -1,11 +1,18 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-data = np.load('bm-R2.npz')
 
-print(data["SIZES"])
+import os
+my_path = os.path.dirname(__file__)
+
+output = os.path.join(my_path, "output/")
+
+
+
+data = np.load(os.path.join(my_path, 'bm-R2.npz'))
+
 labels = [str(size) for size in data["SIZES"]]
-print(labels)
+
 
 
 fig = plt.figure()
@@ -18,9 +25,10 @@ for name in ["T_pa", "T_pds", "T_dr", "T_cvx"]:
     label = name[2:]
     plt.errorbar(range(N_size), mean, yerr=std, label=label)
 
+plt.title("Running time")
 plt.xticks(range(N_size), labels)
 plt.legend()
-plt.savefig("bm-R2-times.png")
+plt.savefig(os.path.join(output, "bm-R2-times.png"))
 plt.show()
 
 
@@ -32,13 +40,14 @@ for name in ["L_pa", "L_pds", "L_dr", "L_cvx"]:
     label = name[2:]
     plt.errorbar(range(N_size), mean, yerr=std, label=label)
 
+plt.title("Value of loss function")
 plt.xticks(range(N_size), labels)
 plt.legend()
-plt.savefig("bm-R2-losses.png")
+plt.savefig(os.path.join(output, "bm-R2-losses.png"))
 plt.show()
 
-for name in ["C_pa", "C_pds", "C_dr", "Cs_cvx"]:
-    L = data[name] - data["L_pa"]
+for name in ["C_pa", "C_pds", "C_dr", "C_cvx"]:
+    L = data[name]
     N_size, N_data = L.shape
     mean = np.mean(L, axis=1)
     std = np.std(L, axis=1)/np.sqrt(N_data)
@@ -48,5 +57,5 @@ for name in ["C_pa", "C_pds", "C_dr", "Cs_cvx"]:
 plt.title("Value of norm(Cbeta)")
 plt.xticks(range(N_size), labels)
 plt.legend()
-plt.savefig("bm-R2-constraint.png")
+plt.savefig(os.path.join(output, "bm-R2-constraint.png"))
 plt.show()
