+++ benchmark/Makefile
@@ -1,9 +1,13 @@
 # Minimal makefile for benchmark
 #
 
-python3.9 R1/bm-R1.py
-python3.9 R1/bm-R1-plot.py
-
-python3.9 R2/bm-R2.py
-python3.9 R2/bm-R2-plot.py
+benchmark:
+	python3.9 R1/bm-R1.py
+	python3.9 R1/bm-R1-plot.py
+	python3.9 R2/bm-R2.py
+	python3.9 R2/bm-R2-plot.py
+	python3.9 C1/bm-C1.py
+	python3.9 C1/bm-C1-plot.py
+	python3.9 C2/bm-C2.py
+	python3.9 C2/bm-C2-plot.py
 

[…]
