+++ benchmark/R2/bm-R2.py
@@ -17,8 +17,7 @@ def huber(r, rho):
     
     
 
-def loss(X, y, lam, rho, beta):
-    lamb = lam*2*LA.norm(X.T.dot(y),np.infty)
+def loss(X, y, lamb, rho, beta):
     r = X.dot(beta) - y
     return np.sum(huber(r,rho)) + lamb*np.sum(abs(beta))
 
@@ -28,18 +27,18 @@ sigma = 0.5
 lam = 0.1
 
 N_per_data = 5
-N_data = 10
-SIZES = [
-    (50, 100),
-    (100, 100),
-    (100, 200),
-    (200, 200)
-]
+N_data = 20
+S = [50, 100, 200, 500, 1000]
+S = [20, 50]
 
+
+SIZES = []
+for i in range(len(S)):
+    SIZES.append((S[i], S[i]))
+    if i+1<len(S):
+        SIZES.append((S[i], S[i+1]))
 
 N_sizes = len(SIZES)
-
-
 T_pa = np.zeros((N_sizes, N_data))
 L_pa = np.zeros((N_sizes, N_data))
 C_pa = np.zeros((N_sizes, N_data))
@@ -66,7 +65,8 @@ for s in range(N_sizes):
     for i in range(N_data):
         (X, C, y), sol = random_data(m, d, d_nonzero, 1, sigma, zerosum=True, seed=i)
         rho = 1.345 * np.sqrt(np.mean(y**2))
-        lamb = lam*2*LA.norm(X.T.dot(y),np.infty)
+        F = abs(y) >= rho 
+        lamb = lam*2*LA.norm(X[F].T.dot(y[F]),np.infty)
 
         t0 = time()
         # classo Path-Alg
@@ -79,8 +79,8 @@ for s in range(N_sizes):
             problem.formulation.rho = rho
             problem.model_selection.StabSel = False
             problem.model_selection.LAMfixed = True
-            problem.model_selection.LAMfixedparameters.rescaled_lam = True
-            problem.model_selection.LAMfixedparameters.lam = lam
+            problem.model_selection.LAMfixedparameters.rescaled_lam = False
+            problem.model_selection.LAMfixedparameters.lam = lamb
             problem.model_selection.LAMfixedparameters.numerical_method = 'Path-Alg'
             problem.solve()
             b_pa.append(problem.solution.LAMfixed.beta)
@@ -97,8 +97,8 @@ for s in range(N_sizes):
             problem.formulation.rho = rho
             problem.model_selection.StabSel = False
             problem.model_selection.LAMfixed = True
-            problem.model_selection.LAMfixedparameters.rescaled_lam = True
-            problem.model_selection.LAMfixedparameters.lam = lam
+            problem.model_selection.LAMfixedparameters.rescaled_lam = False
+            problem.model_selection.LAMfixedparameters.lam = lamb
             problem.model_selection.LAMfixedparameters.numerical_method = 'P-PDS'
             problem.solve()
             b_pds.append(problem.solution.LAMfixed.beta)
@@ -115,8 +115,8 @@ for s in range(N_sizes):
             problem.formulation.rho = rho
             problem.model_selection.StabSel = False
             problem.model_selection.LAMfixed = True
-            problem.model_selection.LAMfixedparameters.rescaled_lam = True
-            problem.model_selection.LAMfixedparameters.lam = lam
+            problem.model_selection.LAMfixedparameters.rescaled_lam = False
[…]
