+++ benchmark/R2/bm-R2.py
@@ -28,8 +28,7 @@ lam = 0.1
 
 N_per_data = 5
 N_data = 20
-S = [50, 100, 200, 500, 1000]
-S = [20, 50]
+S = [50, 100, 200, 500]
 
 
 SIZES = []
