+++ benchmark/C1/bm-C1.py
@@ -21,8 +21,7 @@ lam = 0.1
 
 N_per_data = 5
 N_data = 20
-S = [50, 100, 200, 500, 1000]
-S = [20, 50]
+S = [50, 100, 200, 500]
 
 
 SIZES = []
