+++ benchmark/README.md
@@ -1,0 +1,206 @@
+<img src="https://i.imgur.com/2nGwlux.png" alt="c-lasso" height="150" align="right"/>
+
+# Numerical benchmarks for c-lasso 
+
+
+Here, we summarize numerical benchmarks for the [c-lasso package](https://c-lasso.readthedocs.io/en/latest/) in comparison to cvx.
+
+## Table of Contents
+
+* [Benchmark setup](#installation)
+* [Regression and classification problems](#regression-and-classification-problems)
+* [Results](#getting-started)
+* [Optimization schemes](#optimization-schemes)
+
+* [References](#references)
+
+
+##  Benchmark set-up
+
+###  Tested Regression and classification problems
+
+#### [R1] Standard constrained Lasso regression:             
+
+<img src="https://latex.codecogs.com/gif.latex?\arg\min_{\beta\in&space;R^d}&space;||&space;X\beta-y&space;||^2&space;&plus;&space;\lambda&space;||\beta||_1&space;\qquad\mbox{s.t.}\qquad&space;C\beta=0" />
+
+This is the standard Lasso problem with linear equality constraints on the &beta; vector. 
+The objective function combines Least-Squares for model fitting with l1 penalty for sparsity.   
+
+#### [R2] Contrained sparse Huber regression:                   
+
+<img src="https://latex.codecogs.com/gif.latex?\arg\min_{\beta\in&space;R^d}&space;h_{\rho}(X\beta-y&space;)&space;&plus;&space;\lambda&space;||\beta||_1&space;\qquad\mbox{s.t.}\qquad&space;C\beta=0" />
+
+This regression problem uses the [Huber loss](https://en.wikipedia.org/wiki/Huber_loss) as objective function 
+for robust model fitting with l1 and linear equality constraints on the &beta; vector. The parameter &rho;=1.345.
+
+#### [C1] Contrained sparse classification with Square Hinge loss (Square-Hinge SVM): 
+
+<img src="https://latex.codecogs.com/gif.latex?\arg&space;\min_{\beta&space;\in&space;\mathbb{R}^d}&space;\sum_{i=1}^n&space;l(y_i&space;x_i^\top&space;\beta)&space;&plus;&space;\lambda&space;\left\lVert&space;\beta\right\rVert_1&space;\qquad&space;s.t.&space;\qquad&space;C\beta&space;=&space;0" title="\arg \min_{\beta \in \mathbb{R}^d} \sum_{i=1}^n l(y_i x_i \beta) + \lambda \left\lVert \beta\right\rVert_1 \qquad s.t. \qquad C\beta = 0" />
+
+where the x<sub>i</sub> are the rows of X and l is defined as:
+
+<img src="https://latex.codecogs.com/gif.latex?l(r)&space;=&space;\begin{cases}&space;(1-r)^2&space;&&space;if&space;\quad&space;r&space;\leq&space;1&space;\\&space;0&space;&if&space;\quad&space;r&space;\geq&space;1&space;\end{cases}" title="l(r) = \begin{cases} (1-r)^2 & if \quad r \leq 1 \\ 0 &if \quad r \geq 1 \end{cases}" />
+
+This formulation is similar to [R1] but adapted for classification tasks using the Square Hinge loss
+with (constrained) sparse &beta; vector estimation.
+
+#### [C2] Contrained sparse classification with Huberized Square Hinge loss (Huberized Square-Hinge SVM):        
+
[…]
