+++ benchmark/README.md
@@ -71,80 +71,6 @@ with (constrained) sparse &beta; vector estimation.
 
 
 
-<!---
-<img src="https://i.imgur.com/8tFmM8T.png" alt="Central Park Soil Microbiome" height="250" align="right"/>
-#### pH prediction using the Central Park soil dataset 
-The next microbiome example considers the [Central Park Soil dataset](./examples/pH_data) from [Ramirez et al.](https://royalsocietypublishing.org/doi/full/10.1098/rspb.2014.1988). The sample locations are shown in the Figure on the right.)
--->
-
-#### pH prediction using the 88 soils dataset
-
-The next microbiome example considers the [88 soils dataset](./examples/pH_data) from [Lauber et al., 2009](https://pubmed.ncbi.nlm.nih.gov/19502440/).
-
-The task is to predict pH concentration in the soil from microbial abundance data. A similar analysis is available
-in [Tree-Aggregated Predictive Modeling of Microbiome Data](https://www.biorxiv.org/content/10.1101/2020.09.01.277632v1) 
-with Central Park soil data from [Ramirez et al.](https://royalsocietypublishing.org/doi/full/10.1098/rspb.2014.1988).
-
-Code to run this application is available in [the accompanying notebook](./examples/example-notebook.ipynb) under `pH data`. Below is a summary of a c-lasso problem instance (using the R3 formulation).
- 
-```
-FORMULATION: R3
- 
-MODEL SELECTION COMPUTED:  
-     Lambda fixed
-     Path
-     Stability selection
- 
-LAMBDA FIXED PARAMETERS: 
-     numerical_method = Path-Alg
-     rescaled lam : True
-     threshold = 0.004
-     lam : theoretical
-     theoretical_lam = 0.2182
- 
-PATH PARAMETERS: 
-     numerical_method : Path-Alg
-     lamin = 0.001
-     Nlam = 80
- 
- 
-STABILITY SELECTION PARAMETERS: 
-     numerical_method : Path-Alg
-     method : lam
-     B = 50
-     q = 10
-     percent_nS = 0.5
-     threshold = 0.7
-     lam = theoretical
-     theoretical_lam = 0.3085
-```
-
-The c-lasso estimation results are summarized below:
-
-```
-LAMBDA FIXED : 
-   Sigma  =  0.198
-   Selected variables :  14    18    19    39    43    57    62    85    93    94    104    107    
-   Running time :  0.008s
-
- PATH COMPUTATION : 
-   Running time :  0.12s
-
- STABILITY SELECTION : 
-   Selected variables :  2    12    15    
-   Running time :  0.287s
-```
-
-![Ex4.1](https://github.com/Leo-Simpson/c-lasso/blob/master/figures/examplePH/R3-Beta-path.png)
-
-![Ex4.2](https://github.com/Leo-Simpson/c-lasso/blob/master/figures/examplePH/R3-Sigma-path.png)
-
-![Ex4.3](https://github.com/Leo-Simpson/c-lasso/blob/master/figures/examplePH/R3-StabSel.png)
-
-![Ex4.4](https://github.com/Leo-Simpson/c-lasso/blob/master/figures/examplePH/R3-StabSel-beta.png)
-
-![Ex4.5](https://github.com/Leo-Simpson/c-lasso/blob/master/figures/examplePH/R3-beta.png)
-
 
 ## Optimization schemes
 
@@ -169,19 +95,15 @@ via projection. In the absence of a linear constraint, the method reduces to FB.
 This method can solve problem [R1]. For the Huber problem [R3], 
 P-PDS can solve the mean-shift formulation of the problem (see [6]).
 
-### Projection-free primal-dual splitting method (PF-PDS):
[…]
