+++ benchmark/README.md
@@ -45,31 +45,23 @@ where the x<sub>i</sub> are the rows of X and l is defined as:
 This formulation is similar to [R1] but adapted for classification tasks using the Square Hinge loss
 with (constrained) sparse &beta; vector estimation.
 
-#### [C2] Contrained sparse classification with Huberized Square Hinge loss (L1-Regularized Huberized Square-Hinge SVM):        
-
-<img src="https://latex.codecogs.com/gif.latex?\arg&space;\min_{\beta&space;\in&space;\mathbb{R}^d}&space;\sum_{i=1}^n&space;l_{\rho}(y_i&space;x_i^\top\beta)&space;&plus;&space;\lambda&space;\left\lVert&space;\beta\right\rVert_1&space;\qquad&space;s.t.&space;\qquad&space;C\beta&space;=&space;0" title="\arg \min_{\beta \in \mathbb{R}^d} \sum_{i=1}^n l_{\rho}(y_i x_i\beta) + \lambda \left\lVert \beta\right\rVert_1 \qquad s.t. \qquad C\beta = 0" />
-
-where the x<sub>i</sub> are the rows of X and l<sub>ρ</sub> is defined as:
-
-<img src="https://latex.codecogs.com/gif.latex?l_{\rho}(r)&space;=&space;\begin{cases}&space;(1-r)^2&space;&if&space;\quad&space;\rho&space;\leq&space;r&space;\leq&space;1&space;\\&space;(1-\rho)(1&plus;\rho-2r)&space;&&space;if&space;\quad&space;r&space;\leq&space;\rho&space;\\&space;0&space;&if&space;\quad&space;r&space;\geq&space;1&space;\end{cases}" title="l_{\rho}(r) = \begin{cases} (1-r)^2 &if \quad \rho \leq r \leq 1 \\ (1-\rho)(1+\rho-2r) & if \quad r \leq \rho \\ 0 &if \quad r \geq 1 \end{cases}" />
-
-
-This formulation is similar to [C1] but uses the Huberized Square Hinge loss for robust classification 
-with (constrained) sparse &beta; vector estimation.
-
-
 ## Results
 
 #### R1
 
+Run times for R1. 
 
 ![Run times on R1](./output/bm-R1-times.png)
 
 #### R2
 
+Run times for R2. 
+
 ![Run times on R2](./output/bm-R2-times.png)
 
 #### C1
+
+Run times for C1. 
 
 ![Run times on C1](./output/bm-C1-times.png)
 
