+++ benchmark/README.md
@@ -47,7 +47,7 @@ with (constrained) sparse &beta; vector estimation.
 
 ###  Synthetic data generation and test problem set-ups 
 
-We use the `random_data` function in c-lasso to generate $X$ and $y$. We use the standard `zeroSum` constraint. We vary the number of samples n and dimensionionality d of the problems. The regularization parameter is set to $\lambda$ values...
+We use the `random_data` function in c-lasso to generate X and y. We use the standard `zeroSum` constraint. We vary the number of samples n and dimensionionality d of the problems. The regularization parameter is set to &lambda values...
 
 ## Results
 
