### 2020-12-16 leo-simpson/classo#6 @muellsen
We ran a comprehensive run time comparison now... Currently in the process of providing the result in a presentable format (and including it in the paper); sneak peak here: https://github.com/Leo-Simpson/c-lasso/tree/master/benchmark
