### 2020-12-16 leo-simpson/classo#6 @jbytecode
@muellsen does your output underpin the performance claims in the paper?

Will you make a change in the paper? Please let me know, so I can re-compile it in the review page.
