### 2020-12-30 #2844 @mjsottile
:wave: @glemaitre and @jbytecode : I've been tracking progress in the issues linked above, but in the interest of being able to respond in a timely manner once reviews are complete, please add a comment to this issue thread when you believe you are finished and indicate whether or not you support accepting the submission (and if not, what you believe must be addressed).

Please let me know if you have any questions/concerns as well.  Thanks, and Happy New Year!  :partying_face:
