### 2020-12-30 #2844 @jbytecode
Dear editor @mjsottile,

In my reviewing process, I opened many issues indicating the problems by referencing the corresponding reviewing items, and the authors corrected them. I generally respond with a message "that seems okay" before closing the issue. Please tell me if there is a specific problem with this and let me correct.

A single issue is waiting for correction, the authors wrote a micro benchmark tool and generated some graphics. In the [issue](https://github.com/Leo-Simpson/c-lasso/issues/6) they reported that they would present the results in the paper. I am currently waiting them to add these reports.

Happy new year! Thank you very much for your great efforts in JOSS!
