+++ paper/paper.md
@@ -269,7 +269,7 @@ c-lasso is a fast implementation of the constraint lasso problem. In order to un
 
 Notably, it shows that for high dimensional synthetic data, our three main algorithms for [standard constrained lasso regression](#R1) are faster than cvx algorithm, and give a better a precision. Typically, for $500$ features and $500$ samples, our algorithm *Path-Alg* will be $70$ times faster than the solution provided by cvx.
 
-![Running time and value of the objective function for solutions found with *Path-Alg* (blue), *P-PDS* (yellow), *DR* (green) and cvx (red) for a fixed $\lambda = 0.1$ 20 different synthetic datasets have been considered for each dimensions, and the algorithms have been computed 5 times for each for more stable results. On the left, one can see the objective function for one solution, substracted by the one found by *Path-Alg*](figures/figure_benchmark.png)
+![Running time and value of the objective function for solutions found with *Path-Alg* (blue), *P-PDS* (yellow), *DR* (green) and cvx (red) for a fixed $\lambda = 0.1$ 20 different synthetic datasets have been considered for each dimensions, and the algorithms have been computed 5 times for each for more stable results. On the left, one can see the objective function for one solution, substracted by the one found by *Path-Alg*.](figures/figure_benchmark.png)
 
 
 
