### 2021-01-05 cfd0c9f
add hardware&software info
