### 2021-01-06 leo-simpson/classo#6 @muellsen
Happy New Year! Here is a heads-up. We updated the Paper.md and included a section on numerical benchmarks. We also finalized a README in the benchmark folder https://github.com/Leo-Simpson/c-lasso/tree/master/benchmark that shows the micro-benchmarks. We think, now our performance statements in the paper are backed up by these experiments now.
