### 2021-01-06 #2844 @muellsen
> Hi @glemaitre - good questions, and thanks for the details @muellsen. I think this issue is worth addressing in the paper to explicitly clarify that, in its current state, the package is not an API-compatible drop in replacement for components of scikit-learn. The issue with API compatibility (or the lack thereof) doesn't detract from the quality of the work itself, just the expectations that users may have. If the paper just adds a simple statement that the code currently is not a straight drop-in replacement in scikit-learn, then potential users won't mistakenly assume that it is. I think such a design for c-lasso would be interesting to look at in the future, but isn't critical for this submission.

@glemaitre @mjsottile I added a line in the Paper.md and the README clarifying that c-lasso is a stand-alone package and not yet compatible with the scikit-learn API. It would be absolutely awesome, however, to work on this!
