### 2021-01-06 #2844 @jbytecode
@muellsen, Please check the pdf

- L128: `1,8` might be `1.8` ?
- L129: `8 Go` might be `8 Gb`?
