### 2021-01-06 leo-simpson/classo#6 @jbytecode
seems okay. Thank you. I am now closing this issue.
