+++ benchmark/README.md
@@ -42,7 +42,7 @@ We use the `random_data` function in c-lasso to generate X and y. We use the sta
 
 ## Results
 
-The running times of the micro-benchmark has been computed using Python 3.9.1 on a laptop `MacBook Air`, operating on macOS high Sierra with the processor `1,8 GHz Intel Core i5`, with memory of `8 Go 1600 MHz DDR3`.
+The running times of the micro-benchmark has been computed using Python 3.9.1 on a laptop `MacBook Air`, operating on macOS high Sierra with the processor `1.8 GHz Intel Core i5`, with memory of `8 Gb 1600 MHz DDR3`.
 
 #### R1
 
