### 2021-01-15 #2844 @jbytecode
Dear editor @mjsottile

- The authors added the missing license entry in the software repository.
- Unit test were missing. The authors added automation tests. Now, the package is automatically tested before being available.
- The paper is well written in my opinion. Beyond this, they corrected all of the typos and code formatting that I addressed.
- There were serious performance claims in the earlier versions of the paper. The authors added some benchmark tools to software repository and they extended the paper with a new section that includes reports of these benchmark studies. I think that is okay.
- The authors also added community guidelines.

Summarizing all, the authors considered all of the issues that I addressed. Personally, I can recommend an acceptance.
