### 2021-01-15 #2844 @jbytecode
> > @muellsen, @Leo-Simpson, Please check the pdf
> >
> > * L128: `1,8` might be `1.8` ?
> > * L129: `8 Go` might be `8 Gb`?
>
> Just saw this, sorry. Yes. Updated!

Thank you @muellsen, I am done with this. My review report is above.
