### 2021-01-16 #2844 @mjsottile
Hi @muellsen @Leo-Simpson - I am making my final editorial pass over the text.  A couple questions about the text:

- I noticed a number of places where the word "constrained" is spelled "contrained".  I am assuming this is a typo that you can search and replace to fix.
- In the toy example, it would be useful to reformat slightly so that the closing paren for the random_data() call isn't on its own at the top of the following page.
- Remove capitalization of "laptop" at the end of the section on the toy example.
- Capitalize "Lipschitzian" in the references since it derives from a proper noun.

Once those edits are complete, I believe there will be no remaining outstanding issues and we can proceed to finalizing the paper.  The final steps will be:

- Submit a version of the software to an archive (e.g., figshare or zenodo) and post the DOI for the archive in this thread.

- Please verify that the version number matches that of the archived code, and if not, let me know so I can update the version here.

- Finally, make sure that the metadata for the archive (e.g., title, authors) matches the paper.

Please comment here when you've completed everything and have a DOI.
