### 2021-01-16 de89cb7
Delete README-for-pypi.md
