+++ README-for-pypi.md
@@ -1,267 +1,0 @@
-# c-lasso: a Python package for constrained sparse regression and classification 
-=========
-
-c-lasso is a Python package that enables sparse and robust linear regression and classification with linear equality
-constraints on the model parameters. The forward model is assumed to be: 
-
-<img src="https://latex.codecogs.com/gif.latex?y=X\beta&plus;\sigma\epsilon\qquad\text{s.t.}\qquad&space;C\beta=0" title="y=X\beta+\sigma\epsilon\qquad\text{s.t.}\qquad C\beta=0" />
-
-Here, y and X are given outcome and predictor data. The vector y can be continuous (for regression) or binary (for classification). C is a general constraint matrix. The vector &beta; comprises the unknown coefficients and &sigma; an 
-unknown scale.
-
-The package handles several different estimators for inferring &beta; (and &sigma;), including 
-the constrained Lasso, the constrained scaled Lasso, and sparse Huber M-estimation with linear equality constraints.
-Several different algorithmic strategies, including path and proximal splitting algorithms, are implemented to solve 
-the underlying convex optimization problems.
-
-We also include two model selection strategies for determining the sparsity of the model parameters: k-fold cross-validation and stability selection.   
-
-This package is intended to fill the gap between popular python tools such as [scikit-learn](https://scikit-learn.org/stable/) which CANNOT solve sparse constrained problems and general-purpose optimization solvers that do not scale well for the considered problems.
-
-Below we show several use cases of the package, including an application of sparse *log-contrast*
-regression tasks for *compositional* microbiome data.
-
-The code builds on results from several papers which can be found in the [References](#references).
-
-## Table of Contents
-
-* [Installation](#installation)
-* [Regression and classification problems](#regression-and-classification-problems)
-* [Getting started](#getting-started)
-* [Optimization schemes](#optimization-schemes)
-* [References](#references)
-
-
-##  Installation
-
-c-lasso is available on pip. You can install the package
-in the shell using
-
-```shell
-pip install c_lasso
-```
-To use the c-lasso package in Python, type 
-
-```python
-from classo import *
-```
-
-The c-lasso package depends on several standard Python packages. 
-To import these packages, use 
-
-```shell
-pip install numpy
-pip install matplotlib
-pip install scipy
-pip install pandas
-pip install time
-```
-
-##  Regression and classification problems
-
-The c-lasso package can solve six different types of estimation problems: 
-four regression-type and two classification-type formulations.
-
-#### [R1] Standard constrained Lasso regression:             
-
-<img src="https://latex.codecogs.com/gif.latex?\arg\min_{\beta\in&space;R^d}&space;||&space;X\beta-y&space;||^2&space;&plus;&space;\lambda&space;||\beta||_1&space;\qquad\mbox{s.t.}\qquad&space;C\beta=0" />
-
-This is the standard Lasso problem with linear equality constraints on the &beta; vector. 
-The objective function combines Least-Squares for model fitting with l1 penalty for sparsity.   
-
-#### [R2] Contrained sparse Huber regression:                   
-
-<img src="https://latex.codecogs.com/gif.latex?\arg\min_{\beta\in&space;R^d}&space;h_{\rho}(X\beta-y&space;)&space;&plus;&space;\lambda&space;||\beta||_1&space;\qquad\mbox{s.t.}\qquad&space;C\beta=0" />
-
[…]
