### 2021-01-16 #2844 @Leo-Simpson
Dear @mjsottile,

Thank you very much.

I finished those edits, except for the problem in the references, which I think can not be fixed on the paper, because the references are generated automatically with pandoc with a certain citation rendering, and I assume that it restricts the rendering of the title, because there is a capital letter in the `paper.bib`file.

I submitted the software to figshare : https://doi.org/10.6084/m9.figshare.13589585.v1

So I guess the version of the archived code is 'v1', but I don't know if we can modify it.

The DOI is : 10.6084/m9.figshare.13589585

Best regards.
