### 2021-01-16 #2844 @mjsottile
@Leo-Simpson I believe you can use the usual BiBTeX method to preserve case by wrapping the titles (or individual words) in an additional layer of {} braces.  Pandoc is just using LaTeX to render the markdown into PDF, so that should fix the cases.  It looks like some of your bib file entries already have extra braces like this (eg, see the title entry for bien:2020).
