+++ paper/paper.bib
@@ -2,7 +2,7 @@
 
 @article{Aitchison:1984,
     author = {Aitchion, J. and Bacon-Shone, J.},
-    title = {Log contrast models for experiments with mixtures},
+    title = {{Log contrast models for experiments with mixtures}},
     journal = {Biometrika},
     volume = {71},
     number = {2},
@@ -38,7 +38,7 @@ year = {2020}
 	Journal = {Journal of Optimization Theory and Applications},
 	Number = {3},
 	Pages = {907--924},
-	Title = {A Projected Primal--Dual Method for Solving Constrained Monotone Inclusions},
+	Title = {{A Projected Primal--Dual Method for Solving Constrained Monotone Inclusions}},
 	Ty = {JOUR},
 	Url = {https://doi.org/10.1007/s10957-018-1430-2},
 	Volume = {180},
@@ -50,7 +50,7 @@ year = {2020}
   year = {2012},
   month = {06},
   pages = {307--320},
-  title = {Primal-Dual Splitting Algorithm for Solving Inclusions with Mixtures of Composite, Lipschitzian, and Parallel-Sum Type Monotone Operators},
+  title = {{Primal-Dual Splitting Algorithm for Solving Inclusions with Mixtures of Composite, Lipschitzian, and Parallel-Sum Type Monotone Operators}},
   volume = {20},
   journal = {Set-Valued and Variational Analysis},
   doi = {10.1007/s11228-011-0191-y}
@@ -64,7 +64,7 @@ journal = {Electron. J. Statist.},
 number = {1},
 pages = {207--238},
 publisher = {The Institute of Mathematical Statistics and the Bernoulli Society},
-title = {Perspective maximum likelihood-type estimation via proximal decomposition},
+title = {{Perspective maximum likelihood-type estimation via proximal decomposition}},
 url = {https://doi.org/10.1214/19-EJS1662},
 volume = {14},
 year = {2020}
@@ -79,7 +79,7 @@ year = {2020}
     Id = {Combettes2020},
     Isbn = {1867-1772},
     Journal = {Statistics in Biosciences},
-    Title = {Regression Models for Compositional Data: General Log-Contrast Formulations, Proximal Optimization, and Microbiome Data Applications},
+    Title = {{Regression Models for Compositional Data: General Log-Contrast Formulations, Proximal Optimization, and Microbiome Data Applications}},
     Ty = {JOUR},
     Url = {https://doi.org/10.1007/s12561-020-09283-2},
     Year = {2020},
@@ -88,7 +88,7 @@ year = {2020}
 
 @article{Gaines:2018,
     author = {Brian R. Gaines and Juhyun Kim and Hua Zhou},
-    title = {Algorithms for Fitting the Constrained Lasso},
+    title = {{Algorithms for Fitting the Constrained Lasso}},
     journal = {Journal of Computational and Graphical Statistics},
     volume = {27},
     number = {4},
@@ -103,7 +103,7 @@ year = {2020}
 
 @article{James:2020,
     author = {Gareth M. James and Courtney Paulson and Paat Rusmevichientong},
-    title = {Penalized and Constrained Optimization: An Application to High-Dimensional Website Advertising},
+    title = {{Penalized and Constrained Optimization: An Application to High-Dimensional Website Advertising}},
     journal = {Journal of the American Statistical Association},
     volume = {115},
     number = {529},
@@ -117,7 +117,7 @@ year = {2020}
 
 
 @book{Hastie:2009,
-  title={The Elements of Statistical Learning: Data Mining, Inference, and Prediction},
[…]
