### 2021-01-17 #2844 @arfon
@jbytecode, @glemaitre - many thanks for your reviews here and to @mjsottile for editing this submission. JOSS relies upon the volunteer efforts of folks like yourselves and we simply wouldn't be able to do this without you! ✨
